move IPython.consoleapp to jupyter_client
